Sumfile

+++ Sum.java
@@ -2,14 +2,14 @@ public class Sum	//class Sum wi
 {
 	public int sum(double x, int y)	//function to add two integer values 
 	{
-		return(x+y);
+		return((int)(x+y));
 	}
 	public int sum(int x, int y, int z)		//function to add three integer values 
 	{
 		return(x+y+z);
 	}
-	public double sum(int x, double y)	//function to add two double values 
+	public double sum(int x, int y)	//function to add two double values 
 	{
-		return(x+y);
+		return((double)(x+y));
 	}
 }
